requirements.txt
click==7.1.2
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
SQLAlchemy==1.4.15
Werkzeug==2.0.1
psycopg2-binary==2.9.5
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.1
python-dotenv==0.21.0
celery==5.1.0
redis==3.5.3
# click==8.0.1
# Flask==1.1.2
# Flask-SQLAlchemy==2.5.1
# greenlet==1.1.0
# itsdangerous==2.0.1
# Jinja2==3.0.1
# MarkupSafe==2.0.1
# marshmallow==3.12.1
# marshmallow-sqlalchemy==0.25.0
# SQLAlchemy==1.4.15
# Werkzeug==2.2.2
